// minimal/components/custom-breadcrumbs/types.ts
// @mui
import { BreadcrumbsProps } from "@mui/material/Breadcrumbs";

// ----------------------------------------------------------------------

export type BreadcrumbsLinkProps = {
  name?: string;
  href?: string;
  icon?: React.ReactElement;
};

export interface CustomBreadcrumbsProps extends BreadcrumbsProps {
  headingText?: string;
  showHeadingText?: boolean;
  moreLink?: string[];
  activeLast?: boolean;
  action?: React.ReactNode;
  links: BreadcrumbsLinkProps[];
  subNavbar?: React.ReactNode;
  responsiveBreadcrumbs?: boolean;
  onlyLinkable?: boolean;
}